refactor: major cleanup of build infrastructure and unused code
Build Infrastructure:
- Remove dead Demo target referencing non-existent qyl.demo
- Extract GenerateSchemaDdl() in DuckDbGenerator to eliminate DDL duplication
- Extract BuildSingleImage() helper in IDockerBuild to reduce duplication
- Add ParameterPrefix to ICompile and IGenerate for consistency
- Fix ImmutableArray null check in CSharpGenerator.EmitGenAiAttributes
- Rename MyRegex()/MyRegex1() to ImportPathRegex()/ImportStripRegex()
- Clean up IHasSolution and BuildPaths dead directory references

Deleted Unused Code:
- Remove OTelSemconvGenerator and related generated files
- Remove unused TypeSpec domain specs (50+ .tsp files)
- Remove AspNetCore and GrpcService examples
- Remove chatclient and openai example configs
- Remove qyl.mcp.server.tests (empty placeholder project)
- Remove generated GenAiSemconv files from collector

SDK/Configuration:
- Update to ANcpLua.NET.Sdk 1.6.2 from nuget.org
- Fix qyl.mcp missing usings and package references
- Update CPM with required analyzer packages
- Add AL0012 pragmas for deprecated OTel attribute compatibility

Documentation:
- Add README.md
- Add/update CLAUDE.md files for protocol, tests, eng
- Remove redundant test documentation files

+++ src/qyl.collector/Ingestion/OtlpConverter.cs
@@ -199,8 +199,10 @@ public static class OtlpConverter
         string serviceName,
         Dictionary<string, string> attributes)
     {
-        var startTime = DateTime.UnixEpoch.AddTicks(span.StartTimeUnixNano / 100);
-        var endTime = DateTime.UnixEpoch.AddTicks(span.EndTimeUnixNano / 100);
+        // Convert nanoseconds to DateTime (ulong nanoseconds -> ticks, 1 tick = 100ns)
+        // Cast to long is safe for DateTime calculation (overflow only in year 30,000+)
+        var startTime = DateTime.UnixEpoch.AddTicks((long)(span.StartTimeUnixNano / 100));
+        var endTime = DateTime.UnixEpoch.AddTicks((long)(span.EndTimeUnixNano / 100));
 
         // Extract gen_ai attributes with schema normalization (fallback to deprecated names)
         var (providerName, requestModel, tokensIn, tokensOut) = ExtractGenAiAttributes(attributes);
